Back-End-main/pyproject.toml
[tool.poetry]
name = "chatapp-backend"
version = "0.1.0"
description = "Chatapp backend"
authors = ["jaewonE <importjaewonE@gmail.com>"]
license = "GPL-3.0"
readme = "README.md"

[tool.poetry.dependencies]
python = "3.11.10"
fastapi = "^0.110.0"
uvicorn = "^0.28.0"
pydantic = "^2.6.4"
sqlalchemy = "^2.0.28"
python-dotenv = "^1.0.1"
pandas = "^2.2.1"
python-decouple = "^3.8"
pymysql = "^1.1.0"
python-multipart = "^0.0.9"
email-validator = "^2.1.1"
pyjwt = "^2.8.0"
bcrypt = "^4.2.0"
uuid = "^1.30"
pillow = "^11.0.0"
torch = "^2.5.1"
torchvision = "^0.20.1"
torchaudio = "^2.5.1"
transformers = "4.43.4"
tensorflow = "2.15.1"
sentence-transformers = "^3.2.1"
langchain-experimental = "^0.3.3"
langchain = "^0.3.7"
langchain-community = "^0.3.5"
langchain-core = "^0.3.15"
langchain-text-splitters = "^0.3.2"
langchain-openai = "^0.2.6"
faiss-cpu = "^1.9.0"
bs4 = "^0.0.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
